final commit supongo

+++ Kernel/include/beepDriver.h
@@ -2,8 +2,8 @@
 #define _BEEP_DRIVER_H
 #include <stdint.h>
 
-extern void outb(uint8_t arg1, uint8_t arg2);
-extern uint8_t inb(uint8_t arg);
+extern void outb(uint16_t arg1, uint16_t arg2);
+extern uint16_t inb(uint16_t arg);
 
 void beep(uint32_t freq, int wait);
 void noSound(void);
